pyproject.toml
[tool.poetry]
name = "kurs"
version = "0.1.0"
description = ""
authors = ["dima <zik871@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
requests = "^2.32.3"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
flake8 = "^7.1.2"
isort = "^6.0.0"
black = "^25.1.0"
mypy = "^1.15.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
.git
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'